Move FTPS-specific fixtures to its own test file
Avoid some premature checking of envvars that haven't been set yet
